feat(#1739): Add in added elemental damage affixes for maps
