+++ Common/Data/Affixes/MapAffixes.json
@@ -174,6 +174,45 @@
 		]
 	},
 	{
+		"affixType": "MapFireAddedDamageAffix",
+		"equipTypes": "Map",
+		"tiers": [
+			{
+				"minValue": 20,
+				"maxValue": 30,
+				"minimumLevel": 45,
+				"weight": 1,
+				"strength": 18
+			}
+		]
+	},
+	{
+		"affixType": "MapColdAddedDamageAffix",
+		"equipTypes": "Map",
+		"tiers": [
+			{
+				"minValue": 20,
+				"maxValue": 30,
+				"minimumLevel": 45,
+				"weight": 1,
+				"strength": 18
+			}
+		]
+	},
+	{
+		"affixType": "MapLightningAddedDamageAffix",
+		"equipTypes": "Map",
+		"tiers": [
+			{
+				"minValue": 20,
+				"maxValue": 30,
+				"minimumLevel": 45,
+				"weight": 1,
+				"strength": 18
+			}
+		]
+	},
+	{
 		"affixType": "MapFireConversionAffix",
 		"equipTypes": "Map",
 		"tiers": [
