+++ Common/Subworlds/MappingNPC.cs
@@ -1,4 +1,4 @@
-﻿using PathOfTerraria.Common.Systems.Affixes.ItemTypes;
+using PathOfTerraria.Common.Systems.Affixes.Maps;
 using PathOfTerraria.Common.Systems.MobSystem;
 using PathOfTerraria.Core.Items;
 using SubworldLibrary;
