+++ Common/Systems/Affixes/Maps/MapStealFocusChargeAffix.cs
@@ -1,0 +1,11 @@
+using PathOfTerraria.Common.Systems.Charges;
+
+namespace PathOfTerraria.Common.Systems.Affixes.Maps;
+
+public class MapStealFocusChargeAffix : MapAffix
+{
+	public override void OnHitPlayer(NPC npc, Player player, Player.HurtInfo info)
+	{
+		player.GetModPlayer<FocusChargePlayer>().RemoveCharge();
+	}
+}
