+++ Common/Systems/Synchronization/Handlers/SendMappingDomainInfoHandler.cs
@@ -1,6 +1,6 @@
 using PathOfTerraria.Common.Subworlds;
 using PathOfTerraria.Common.Systems.Affixes;
-using PathOfTerraria.Common.Systems.Affixes.ItemTypes;
+using PathOfTerraria.Common.Systems.Affixes.Maps;
 using System.Collections.Generic;
 using System.IO;
