+++ Content/Items/Consumables/Maps/BossMaps/EoLMap.cs
@@ -1,6 +1,6 @@
-﻿using PathOfTerraria.Common.Subworlds.BossDomains.Hardmode;
+using PathOfTerraria.Common.Subworlds.BossDomains.Hardmode;
 using PathOfTerraria.Common.Systems.Affixes;
-using PathOfTerraria.Common.Systems.Affixes.ItemTypes;
+using PathOfTerraria.Common.Systems.Affixes.Maps;
 using PathOfTerraria.Content.Items.Currency;
 using PathOfTerraria.Core.Items;
 using SubworldLibrary;
